% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rosetta.R
\name{ann_predict}
\alias{ann_predict}
\title{Get Artificial Neural Network Predictions and Statistics}
\usage{
ann_predict(object, soildata, sum_data)
}
\arguments{
\item{object}{\emph{Rosetta} object containing class instance (e.g. from \code{Rosetta()})}

\item{soildata}{A list containing vectors; with number of parameters matching the model type of \code{object}}

\item{sum_data}{Default: \code{TRUE}}
}
\description{
Get Artificial Neural Network Predictions and Statistics
}
\examples{
# ann_predict(Rosetta(), list(c(30, 30, 40, 1.5), c(55, 25, 20, 1.1)))
}
