% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_python.R
\name{find_python}
\alias{find_python}
\title{Heuristics to Find Python}
\usage{
find_python(envname = NULL, arcpy_path = getOption("rosettaPTF.arcpy_path"))
}
\arguments{
\item{envname}{Default: \code{NULL}; Passed as a preference in \code{reticulate::py_discover_config()} As in \code{reticulate::py_install()}:    The name, or full path, of the environment in which Python packages are to be installed. When NULL (the default), the active environment as set by the RETICULATE_PYTHON_ENV variable will be used; if that is unset, then the r-reticulate environment will be used.}

\item{arcpy_path}{Optional: Path to ArcGIS Pro Python installation. For example: \code{"C:/Program Files/ArcGIS/Pro/bin/Python"}. Set as \code{NULL} to prevent use of ArcGIS Pro instance.}
}
\value{
character path to \code{python} or \code{NULL} if no suitable environment can be found. The result is stored as the package option \code{rosettaPTF.python_path}.
}
\description{
If you are using the {rosettaPTF} package for the first time you will need to have Python installed to obtain the necessary modules. You can set up {reticulate} to install into a virtual or Conda environment. Usually {reticulate} should cover most or all of the setup.
}
\details{
If you have Python set up correctly you should be able to run \code{reticulate::py_config()} or \code{reticulate::py_discover_config()} and discover your shared library.

A common problem with {reticulate} is not pointing at the correct (or any) \code{python} binary or \code{libpython} shared library. Use \code{reticulate::use_python("/path/to/python", required = TRUE)} to set the path or, alternately, be sure \code{python} can be found on your PATH. {reticulate} has a preference for Python environments that have \code{numpy} installed.
\subsection{Windows / Miniconda}{

Use \code{reticulate::install_miniconda()} if you'd like to install a Miniconda Python environment. Conda is default on Windows.

For devices with limited ability to install new software that have ArcGIS Pro installed (some USDA computers), this method can look for a Python installation in \code{"C:/Program Files/ArcGIS/Pro/bin/Python/envs/arcgispro-py3"} and Conda executable in \code{"C:/Program Files/ArcGIS/Pro/bin/Python/Scripts"}. The base file path to "Python" directory can be customized using the \code{arcpy_path} argument.
}

\subsection{Linux and OS X}{

On Linux and OS X you can create a virtual environment using \code{reticulate::virtualenv_create()}. The default environment name will be \code{"r-reticulate"}.
}
}
\examples{

find_python()

}
