% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosetta.R
\name{get_rosetta_module}
\alias{get_rosetta_module}
\title{Get rosetta instance}
\usage{
get_rosetta_module()
}
\value{
An R object wrapping a Python module. Module attributes can be accessed via the \code{$} operator, or via \code{py_get_attr()}.
}
\description{
Get rosetta instance
}
