% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{MUKEY_WCS}
\alias{MUKEY_WCS}
\title{Sample SSURGO Mapunit Web Coverage Service Data: \emph{matrix} of \code{mukey}}
\description{
Sample SSURGO Mapunit Web Coverage Service Data: \emph{matrix} of \code{mukey}
}
