% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rosetta.R
\name{predict.UnsaturatedK}
\alias{predict.UnsaturatedK}
\title{Predict K0 and L from retention parameters}
\usage{
\method{predict}{UnsaturatedK}(object, retc_params, ...)
}
\arguments{
\item{object}{\emph{UnsaturatedK} object}

\item{retc_params}{A list or matrix of retention parameters (theta_r, theta_s, alpha, npar)}

\item{...}{not used}
}
\value{
a \code{data.frame} with \code{log10_K0_mean}, \code{lpar_mean}, \code{log10_K0_sd}, \code{lpar_sd}
}
\description{
Predict K0 and L from retention parameters
}
