% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rosetta.R
\name{Rosetta}
\alias{Rosetta}
\title{Make a Rosetta object instance for running \code{predict()} methods}
\usage{
Rosetta(rosetta_version = 3, model_code = 3)
}
\arguments{
\item{rosetta_version}{Default: \code{3}}

\item{model_code}{One of \code{2}, \code{3}, \code{4}, \code{5}, or \code{-1}. Corresponding to options described in \emph{Details}.}
}
\value{
an instance of the \code{Rosetta} class defined by the Python module; suitable for running \code{predict} or \code{ann_predict} methods.
}
\description{
Make a Rosetta object instance for running \code{predict()} methods
}
\details{
\subsection{Explanation of Model Codes}{
\itemize{
\item \code{2}: 	sand, silt, clay ("SSC")
\item \code{3}: 	sand, silt, clay + bulk density ("BD")
\item \code{4}: 	sand, silt, clay + bulk density + field capacity water content (1/3 bar or 33 kPa tension)
\item \code{5}: 	sand, silt, clay + bulk density + field capacity water content + wilting point water content (15 bar or 1500 kPa tension)
\item \code{-1}: 	no result returned, inadequate or erroneous data
}
}
}
