% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_utils.R
\name{rosetta_pkg_version}
\alias{rosetta_pkg_version}
\title{Get rosetta-soil Python package version}
\usage{
rosetta_pkg_version()
}
\value{
\code{package_version} object
}
\description{
Get rosetta-soil Python package version
}
\keyword{internal}
