% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_utils.R
\name{SoilDataFromArray}
\alias{SoilDataFromArray}
\title{Convert list of numeric vectors to SoilData python object}
\usage{
SoilDataFromArray(x)
}
\arguments{
\item{x}{a list of numeric vectors}
}
\value{
a reference to a Rosetta SoilData python object instance
}
\description{
Convert list of numeric vectors to SoilData python object
}
