% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rosetta.R
\name{UnsaturatedK}
\alias{UnsaturatedK}
\title{Make an UnsaturatedK object instance}
\usage{
UnsaturatedK()
}
\value{
an instance of the \code{UnsaturatedK} class.
}
\description{
\code{UnsaturatedK}: Create an instance of the \code{UnsaturatedK} class from \code{rosetta-soil} >= 0.3. This class is used to predict \code{K0} and \code{L} from retention parameters.
}
