% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Rosetta.R
\name{predict.Rosetta}
\alias{predict.Rosetta}
\title{Predict Rosetta Parameter Values and Standard Deviations from a \emph{Rosetta} instance}
\usage{
\method{predict}{Rosetta}(object, soildata, ...)
}
\arguments{
\item{object}{\emph{Rosetta} object containing class instance (e.g. from \code{Rosetta()})}

\item{soildata}{A list containing vectors; with number of parameters matching the model type of \code{object}}

\item{...}{not used}
}
\description{
Predict Rosetta Parameter Values and Standard Deviations from a \emph{Rosetta} instance
}
\examples{
# predict(Rosetta(), list(c(30, 30, 40, 1.5), c(55, 25, 20, 1.1)))
}
