% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_utils.R
\name{rosetta_module_available}
\alias{rosetta_module_available}
\title{Check if rosetta module can be found in local python environment}
\usage{
rosetta_module_available()
}
\value{
\emph{logical}
}
\description{
Check if rosetta module can be found in local python environment
}
