% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettaPTF-package.R
\docType{package}
\name{rosettaPTF-package}
\alias{rosettaPTF}
\alias{rosettaPTF-package}
\title{rosettaPTF: R Frontend for Rosetta Pedotransfer Functions}
\description{
Access Python rosetta-soil pedotransfer functions in an R environment. Rosetta is a neural network-based model for predicting unsaturated soil hydraulic parameters from basic soil characterization data. The model predicts parameters for the van Genuchten unsaturated soil hydraulic properties model, using sand, silt, and clay, bulk density and water content. The codebase is now maintained by Dr. Todd Skaggs and other U.S. Department of Agriculture employees. This R package is intended to provide for use cases that involve many thousands of calls to the pedotransfer function. Less demanding use cases are encouraged to use the web interface or API endpoint. There are additional wrappers of the API endpoints provided by the soilDB R package `ROSETTA()` method.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ncss-tech/rosettaPTF}
  \item \url{https://ncss-tech.github.io/rosettaPTF/}
  \item Report bugs at \url{https://github.com/ncss-tech/rosettaPTF/issues}
}

}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov}

}
\keyword{internal}
