% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rosetta.R
\name{run_rosetta.default}
\alias{run_rosetta.default}
\alias{run_rosetta}
\alias{run_rosetta.data.frame}
\alias{run_rosetta.RasterStack}
\title{Run \code{rosetta()} method from Python module}
\usage{
\method{run_rosetta}{default}(soildata, vars = NULL, rosetta_version = 3)

\method{run_rosetta}{data.frame}(soildata, vars = NULL, rosetta_version = 3)

\method{run_rosetta}{RasterStack}(soildata, vars = NULL, rosetta_version = 3)
}
\arguments{
\item{soildata}{A list of numeric vectors each containing 3 to 6 values: \code{"sand"}, \code{"silt"}, \code{"clay"}, \code{"bulkdensity"}, \code{"th33"}, \code{"th1500"} or a \emph{data.frame} with 3 to 6 columns.}

\item{vars}{\emph{character}. Optional: names and order of custom column names if \code{soildata} is a \emph{data.frame}. Default \code{NULL} assumes input column order follows \code{sand}, \code{silt}, \code{clay}, \code{bulkdensity}, \code{th33}, \code{th1500} and does not check names.}

\item{rosetta_version}{Default: 3}
}
\value{
A \emph{data.frame} containing \code{mean} and \code{stdev} for following five columns (parameters for van Genuchten-Mualem equation)
\itemize{
\item \code{theta_r}, residual water content
\item \code{theta_s}, saturated water content
\item \code{log10(alpha)}, 'alpha' shape parameter, log10(1/cm)
\item \code{log10(npar)}, 'n' shape parameter
\item \code{log10(Ksat)}, saturated hydraulic conductivity, log10(cm/day)
}
}
\description{
Run \code{rosetta()} method from Python module
}
