% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta.R
\name{run_rosetta}
\alias{run_rosetta}
\title{Run \code{rosetta()} method from Python module}
\usage{
run_rosetta(soildata, rosetta_version = 3)
}
\arguments{
\item{soildata}{A list of numeric vectors containing 3 to 6 elements: \code{"sand"}, \code{"silt"}, \code{"clay"}, \code{"bulkdensity"}, \code{"th33"}, \code{"th1500"}}

\item{rosetta_version}{Default: 3}
}
\value{
A \emph{data.frame} containing \code{mean} and \code{stdev} for following five columns (parameters for van Genuchten-Mualem equation)
\itemize{
\item \code{theta_r}, residual water content
\item \code{theta_s}, saturated water content
\item \code{log10(alpha)}, 'alpha' shape parameter, log10(1/cm)
\item \code{log10(npar)}, 'n' shape parameter
\item \code{log10(Ksat)}, saturated hydraulic conductivity, log10(cm/day)
}
}
\description{
Run \code{rosetta()} method from Python module
}
