% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_utils.R
\name{SoilDataFromArray}
\alias{SoilDataFromArray}
\title{Convert list of numeric vectors to \emph{SoilData} Python object}
\usage{
SoilDataFromArray(x)
}
\arguments{
\item{x}{a list of numeric vectors}
}
\value{
an object reference to a Rosetta \emph{SoilData} Python object constructed from \code{x}
}
\description{
Convert list of numeric vectors to \emph{SoilData} Python object
}
